# specs/002-i-need-to/contracts/api-contracts.yaml
openapi: 3.1.1
info:
  title: Calorie Track AI Bot - Integration & UI Enhancement API
  description: API contracts for backend-frontend integration, connectivity fixes, and UI enhancements
  version: 1.0.0
  contact:
    name: Calorie Track AI Bot Team
    email: support@calorietrack.app
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://calorie-track-ai-bot.fly.dev
    description: Production server
paths:
...
  /api/v1/dev/environment:
    get:
      summary: Get development environment status
      description: Get current development environment configuration with Supabase CLI status
      tags:
        - Development
      responses:
        '200':
          description: Environment status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevelopmentEnvironment'
        '404':
          description: Development environment not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/dev/supabase/status:
    get:
      summary: Get Supabase database status
      description: Check if Supabase local database instance is running
      tags:
        - Development
      responses:
        '200':
          description: Supabase database status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [running, stopped, error]
                  db_url:
                    type: string
                    format: uri
                    description: Database connection URL
                  db_port:
                    type: integer
                    description: Database port (typically 54322)
                  version:
                    type: string
                    description: Supabase CLI version
                  services:
                    type: object
                    properties:
                      db:
                        type: boolean
                        description: Database service status
        '500':
          description: Error checking Supabase database status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DevelopmentEnvironment:
      type: object
      required:
        - id
        - name
        - frontend_port
        - backend_port
        - supabase_local_url
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        frontend_port:
          type: integer
          minimum: 1024
          maximum: 65535
        backend_port:
          type: integer
          minimum: 1024
          maximum: 65535
        supabase_db_url:
          type: string
          format: uri
          description: Supabase database URL (typically http://localhost:54322)
        supabase_db_password:
          type: string
          description: Database password
        redis_url:
          type: string
          format: uri
        storage_endpoint:
          type: string
          format: uri
        cors_origins:
          type: array
          items:
            type: string
            format: uri
        log_level:
          type: string
          enum: [DEBUG, INFO, WARNING, ERROR, CRITICAL]
        hot_reload:
          type: boolean
        supabase_cli_version:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - correlation_id
      properties:
        error:
          type: string
        message:
          type: string
        correlation_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...